Make side menu optionally inline or full page + remove /edit route
